refactor: key reg test utils
Refactors key registration test utils to use real transaction data from the Lora explorer.

This removes the old test implementation of key registration and provides a more complete integration test with real data,
making it easier to test and debug key registration transactions.
